Remove a ton of stale copyright code
It's something like 1700 lines, which seems like a lot, but it's all either
commented or only called from code that's commented out. So it hasn't been
used or tested or much of anything for right about four years.

We're going to be doing something similar shortly, and I hope we can pull
bits from this older code. In the meantime, though, it has to go so we can
start to walk in that direction again.